finished basic evaluation, started writing search

+++ src/utils.c
@@ -6,12 +6,12 @@
 
 #include <stdlib.h>
 
-unsigned int cur_time_millis(void) {
+unsigned long cur_time_millis(void) {
 #ifdef WIN32
     return (unsigned int) GetTickCount();
 #else
     struct timeval timeval;
     gettimeofday(&timeval, NULL);
-    return timeval.tv_sec * 1000 + (timeval.tv_usec / 1000);
+    return (long) timeval.tv_sec * 1000 + (timeval.tv_usec / 1000);
 #endif
 }
